include client filters into TrackedFilteredServiceTest
