+++ Sources/ServiceModel.Grpc.DesignTime.Generator.Test/AspNetCore/TrackedFilteredServiceTest.cs
@@ -34,17 +34,18 @@ public class TrackedFilteredServiceTest : TrackedFilteredServiceTestBase
         _host = await new KestrelHost()
             .ConfigureServices(services =>
             {
-                services.AddTransient<TrackingServerFilter>();
+                services.AddSingleton(new TrackingServerFilter("global-server"));
 
                 services.AddServiceModelGrpc(options =>
                 {
-                    options.Filters.Add(1, _ => new TrackingServerFilter("global"));
+                    options.Filters.Add(1, provider => provider.GetRequiredService<TrackingServerFilter>());
                 });
                 services.AddTrackedFilteredServiceOptions(options =>
                 {
                     options.Filters.Add(2, _ => new TrackingServerFilter("service-options"));
                 });
             })
+            .ConfigureClientFactory(ConfigureClientFactory)
             .ConfigureEndpoints(endpoints =>
             {
                 endpoints.MapTrackedFilteredService();
@@ -52,8 +53,12 @@ public class TrackedFilteredServiceTest : TrackedFilteredServiceTestBase
             .StartAsync()
             .ConfigureAwait(false);
 
-        _host.ClientFactory.AddFilteredServiceClient();
-        DomainService = _host.ClientFactory.CreateClient<IFilteredService>(_host.Channel);
+        _host.ClientFactory.AddFilteredServiceClient(options =>
+        {
+            options.Filters.Add(2, new TrackingClientFilter("client-options"));
+        });
+
+        InitializeClient(_host.ClientFactory, _host.Channel);
     }
 
     [OneTimeTearDown]
